moved useEffect logic in MapInfo page to a custom hook
